+++ src/App.vue
@@ -1,7 +1,9 @@
 <script setup lang="ts">
+import NavigationBar from './components/NavigationBar.vue'
 </script>
 
 <template>
+  <NavigationBar />
   <router-view />
 </template>