+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Add npm package `open-props`.
+- Add the navigation bar component.
 
 ## [0.0.1] - 2022-05-19
 ### Added